+++ samples/lib/widgets/bottom_navigation_bar.dart
@@ -32,7 +32,7 @@ class _DemoState extends State<_Demo> {
   Widget build(BuildContext context) => FBottomNavigationBar(
         index: index,
         onChange: (index) => setState(() => this.index = index),
-        items: [
+        children: [
           FBottomNavigationBarItem(
             icon: FAssets.icons.home,
             label: 'Home',