+++ samples/lib/widgets/scaffold.dart
@@ -97,7 +97,7 @@ class _DemoState extends State<_Demo> {
         footer: FBottomNavigationBar(
           index: index,
           onChange: (index) => setState(() => this.index = index),
-          items: [
+          children: [
             FBottomNavigationBarItem(
               icon: FAssets.icons.home,
               label: 'Home',