+++ forui/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add `FButton.icon(...)`.
 
+* Add `FBottomNavigationBar.of(...)`
+
 * Add `FFormFieldStyle`.
 
 * Add `FResizable.semanticFormatterCallback`.
@@ -15,6 +17,10 @@
 ### Changes
 
 * **Breaking:** Change `FAlertIconStyle.height` to `FAlertIconStyle.size`.
+
+* **Breaking:** Rename `FBottomNavigationBar.items` to `FBottomNavigationBar.children`.
+
+* **Breaking:** Remove `FBottomNavigationBar.raw(...)` - use the default constructor instead.
 
 * Change `FResizable` to resize by `FResizable.resizePercentage` when using a keyboard.
 
@@ -28,7 +34,7 @@
 
 * **Breaking:** Change how `FTextFieldStyle` stores various state-dependent styles.
 
-* **Breaking:** Remove `FTextField.error`.
+* **Breaking:** Remove `FTextField.error` - use `FTextField.forceErrorText` instead.
 
 ### Fixes